Uses explicit values for transducers, adds unbuffered eventstream
--- src/**/explicit-* files ---
There was an attempt to be too damn general with transducers,
where it was getting difficult to implement new ones because
they were designed to be done with a transducer constructor, and
each would return an AwaIterable instance.

Now, transducers are more explicitly implemented, and each provides
a simple interface that accepts a (function, sequence) and returns
an object with an iterator method. The code is easier to read, and
slightly easier to type.

TODOs:
  - Remove AwaIterable class. It is just an interface. The challenge
    is figuring out how to type it. The type coverage so far seems
    ok. How can we be sure?
  - Remove any non-explicit transducers, plus the transducer tests,
    plus any type tests for them.

--- Unbuffered event streams ---
The default `eventStream` now is totally unbuffered. The behavior thus
is such that any event emitted prior to some sink beginning to consume
the stream will be lost. In addition, if there are duplicates of the
event fired in the same tick, only the first will be produced. Lastly,
these have NO end event!

Previous event streams are now exported as `bufferedEventStream`. These
can consume an end event, and have a buffer to prevent some of the
drawbacks described above.

TODOs:
  - Implement end events for unbuffered event streams?
  - Buffered streams should parameterize the buffer size.
  - Type, test, and document the buffers themselves.
  - Document how sending a sentinel will stop the current iterator. If
    you want to close the stream altogether, you would need to keep
    `done` state on the stream, not its iterator.
  - Simplify SENTINEL API.